feat : create luckies inquiry

+++ src/main/java/ongjong/namanmoo/domain/Family.java
@@ -28,7 +28,7 @@ public class Family {
     private String inviteCode;
 
     @Column(columnDefinition = "bigint default 0")
-    private Long challengeFamilyCount;
+    private Integer challengeFamilyCount;
 
     private Long familyOwnerId;
